pour gaby 3

+++ src/main/java/com/rostand/FarmBotWEBv2/DTO/CreateReglagesDTO.java
@@ -1,5 +1,6 @@
 package com.rostand.FarmBotWEBv2.DTO;
 
+import com.rostand.FarmBotWEBv2.Entity.Plante;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,9 +10,9 @@ import javax.persistence.Column;
 @Setter
 public class CreateReglagesDTO {
     private int frequence_scan;
-    private int G1;
-    private int G2;
-    private int G3;
+    private Long G1;
+    private Long G2;
+    private Long G3;
     private int PostRefOutilX;
     private int PostRefOutilY;
     private int PostRefOutilZ;
